Breakdown notification duck.
