+++ src/App.tsx
@@ -28,7 +28,7 @@ import { requestDomains } from 'src/store/domains/domains.actions';
 import { async as imagesAsync } from 'src/store/images';
 import { requestLinodes } from 'src/store/linodes/linodes.actions';
 import { async as typesAsync } from 'src/store/linodeTypes';
-import { requestNotifications } from 'src/store/notifications';
+import { requestNotifications } from 'src/store/notification/notification.requests';
 import { requestProfile } from 'src/store/profile';
 import { requestRegions } from 'src/store/regions/regions.actions';
 import composeState from 'src/utilities/composeState';
